fix: missing revert when pid rollover happened
When there a container belonging to a pod is restarted and get's
assigned a pid which is lower than the pause container we reference the
namespace path which belongs to the container process and not the pause
process, which would be correct. When that happened and the container is
restarted a second time the rollback is skipped because the path
remembered is gone, but the namespace is still there.

To workaround this, we remember now the namespace additionally by it's
inode and will resolve the path when we need it.