fix: revert on missing container
Internally the active attacks are tracked by namespace Inode number.
Currently when a container exits during the attack, we skip the revert
of the attack entirely when that namespace is gone. But this leaks the
namespace inode number in the internal tracking. Which will be a problem
if that inode number will be recycled for a different container. Then
the attack will be blocked by the tracking, although there is no attack
going on. With this change, we will anyway call the revert of the
attack, so that the tracking is correct, but ignore any failures coming
up.